+++ FileDownloaderDemo2/src/main/java/org/wlf/filedownloader_demo2/FileDownloadApplication.java
@@ -61,7 +61,10 @@ public class FileDownloadApplication extends Application {
         builder.configRetryDownloadTimes(5);
 
         // enable debug mode
-        builder.configDebugMode(true);
+        //builder.configDebugMode(true);
+
+        // config connect timeout
+        //builder.configConnectTimeout(25000); // 25s
 
         // 3.init FileDownloader with the configuration
         // build FileDownloadConfiguration with the builder